Setup WMS main to handle moving multiple robots per time Setup
Moves the logic of what steps a robot should do in a task into a
separate task instance. By having a task object with state we can make
the WMS move multiple robots forward one step at a time based on each
robots task and how far it's gotten with that task.

Using a separate task object that knows how far into the
task the robot is, the robot could also pause the task to go for a
battery re-charge break.