Improve charging
On a high level robots now consume a battery charge when moving and can
be told to go to recharge by the ICA algoritm (however the ICA algoritm
crashes for some inputs). Dynamically aborting a mission to go charge
is not yet implemented.

This change moved and attached the task object into the robot class and
created a separate task for charging. In the robot class there is now
two tasks, one normal task for moving packets, and a another task that
will handle charging of the robot. In the case that a robot needs to
charge the charging task in the robot will take precedence. But on the
outside we just call robot.timeStep() and the robot will itself decide
what actions to execute.

In the WMS abortToCharge now handles setting up the robot with a task
for charging. That task should remove itself when complete.

Also fixed a bug in the ICA algorithm that was due to the algoritm
sometimes trying to calculate a path right into a charging station and
not to the loading position besides into. Moving right into a charging
station doesn't work because they are seen as obstacles on the map.